`Options::remove()` now returns `true` when an option does not exist This prevents an error message from being displayed when trying to reset settings before saving them

+++ includes/Options/WP_Options.php
@@ -51,11 +51,17 @@ class WP_Options implements Options {
     /**
      * Remove the option with the given name.
      *
+     * If the option does not exist, this method will return `true`.
+     *
      * @param string $name Option name.
      *
      * @return bool Whether the option was removed successfully.
      */
     public function remove( string $name ): bool {
+        if ( ! $this->has( $name ) ) {
+            return true;
+        }
+
         return delete_option( $name );
     }
